Split `etl-destinations` into multiple subcrates
