+++ Cargo.toml
@@ -6,6 +6,8 @@ members = [
     "etl-benchmarks",
     "etl-config",
     "etl-destinations",
+    "etl-destinations/bigquery",
+    "etl-destinations/iceberg",
     "etl-examples",
     "etl-postgres",
     "etl-replicator",
@@ -17,6 +19,8 @@ etl = { path = "etl", default-features = false }
 etl-api = { path = "etl-api", default-features = false }
 etl-config = { path = "etl-config", default-features = false }
 etl-destinations = { path = "etl-destinations", default-features = false }
+etl-destinations-bigquery = { path = "etl-destinations/bigquery", default-features = false }
+etl-destinations-iceberg = { path = "etl-destinations/iceberg", default-features = false }
 etl-postgres = { path = "etl-postgres", default-features = false }
 etl-replicator = { path = "etl-replicator", default-features = false }
 etl-telemetry = { path = "etl-telemetry", default-features = false }
