+++ etl-destinations/bigquery/src/encryption.rs
