+++ etl-destinations/bigquery/tests/bigquery_pipeline.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "bigquery")]
-
 use chrono::{DateTime, NaiveDate, NaiveDateTime, NaiveTime, Utc};
 use etl::config::BatchConfig;
 use etl::error::ErrorKind;
@@ -10,7 +8,7 @@ use etl::test_utils::pipeline::{create_pipeline, create_pipeline_with};
 use etl::test_utils::test_destination_wrapper::TestDestinationWrapper;
 use etl::test_utils::test_schema::{TableSelection, insert_mock_data, setup_test_database_schema};
 use etl::types::{EventType, PgNumeric, PipelineId};
-use etl_destinations::bigquery::install_crypto_provider_for_bigquery;
+use etl_destinations_bigquery::install_crypto_provider_for_bigquery;
 use etl_telemetry::tracing::init_test_tracing;
 use rand::random;
 use std::str::FromStr;
