+++ etl-destinations/bigquery/tests/support/mod.rs
@@ -1,0 +1,1 @@
+pub mod bigquery;
