+++ etl-destinations/iceberg/tests/iceberg_pipeline.rs
@@ -1,7 +1,5 @@
-#![cfg(feature = "iceberg")]
-
 use etl::types::{ColumnSchema, TableId, TableName, TableSchema, Type};
-use etl_destinations::iceberg::IcebergClient;
+use etl_destinations_iceberg::IcebergClient;
 use etl_telemetry::tracing::init_test_tracing;
 
 use crate::support::lakekeeper::LakekeeperClient;
