+++ etl-destinations/iceberg/tests/support/lakekeeper.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)]
-
 use uuid::Uuid;
 
 pub struct LakekeeperClient {
