+++ etl-destinations/iceberg/tests/support/mod.rs
@@ -1,2 +1,1 @@
-pub mod bigquery;
 pub mod lakekeeper;
